+++ Sources/SuperwallKit/Debug/DebugViewController.swift
@@ -265,7 +265,11 @@ final class DebugViewController: UIViewController {
       return
     }
 
-    let child = factory.makePaywallViewController(for: paywall, withCache: nil)
+    let child = factory.makePaywallViewController(
+      for: paywall,
+      withCache: nil,
+      delegate: nil
+    )
     addChild(child)
     previewContainerView.insertSubview(child.view, at: 0)
     previewViewContent = child.view