+++ Sources/SuperwallKit/Paywall/Presentation/Get PaywallViewController/Operators/MapErrorsOperator.swift
@@ -9,30 +9,30 @@ import Foundation
 import Combine
 
 extension AnyPublisher where Output == PaywallViewController, Failure == Error {
-  func mapErrors(isObjc: Bool) -> AnyPublisher<PaywallViewController, Error> {
+  func mapErrors(toObjc: Bool) -> AnyPublisher<PaywallViewController, Error> {
     mapError { error -> Error in
       if let error = error as? PresentationPipelineError {
         switch error {
         case .holdout(let experiment):
-          if isObjc {
+          if toObjc {
             return PaywallSkippedReasonObjc.holdout
           } else {
             return PaywallSkippedReason.holdout(experiment)
           }
         case .noRuleMatch:
-          if isObjc {
+          if toObjc {
             return PaywallSkippedReasonObjc.noRuleMatch
           } else {
             return PaywallSkippedReason.noRuleMatch
           }
         case .eventNotFound:
-          if isObjc {
+          if toObjc {
             return PaywallSkippedReasonObjc.eventNotFound
           } else {
             return PaywallSkippedReason.eventNotFound
           }
         case .userIsSubscribed:
-          if isObjc {
+          if toObjc {
             return PaywallSkippedReasonObjc.userIsSubscribed
           } else {
             return PaywallSkippedReason.userIsSubscribed