+++ Sources/SuperwallKit/Paywall/Presentation/Get PaywallViewController/PublicGetPaywallViewController.swift
@@ -13,15 +13,6 @@ extension Superwall {
   /// Gets the  ``PaywallViewController`` object for an event, which you can present
   /// however you want.
   ///
-  /// To use this you **must** follow the following steps:
-  ///
-  /// 1. Call this function to retrieve the ``PaywallViewController``.
-  /// 2. Call ``PaywallViewController/presentationWillBegin()`` when
-  /// you're about to present the view controller.
-  /// 3. Present the view controller.
-  /// 4. Call ``PaywallViewController/presentationDidFinish()`` after presentation
-  /// completes.
-  ///
   /// - Note: The remotely configured presentation style will be ignored, it is up to you
   /// to set it programmatically.
   ///
@@ -32,6 +23,7 @@ extension Superwall {
   ///   JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will
   ///   be dropped.
   ///   - paywallOverrides: An optional ``PaywallOverrides`` object whose parameters override the paywall defaults. Use this to override products, presentation style, and whether it ignores the subscription status. Defaults to `nil`.
+  ///   - delegate: A delegate responsible for handling user interactions with the retrieved ``PaywallViewController``.
   ///   - completion: A completion block accepting an optional ``PaywallViewController``, an optional
   ///   ``PaywallSkippedReason`` and an optional `Error`. If the ``PaywallViewController`` couldn't be retrieved
   ///   because its presentation should be skipped, the ``PaywallSkippedReason`` will be non-`nil`. Any errors
@@ -40,6 +32,7 @@ extension Superwall {
     forEvent event: String,
     params: [String: Any]? = nil,
     paywallOverrides: PaywallOverrides? = nil,
+    delegate: PaywallViewControllerDelegate,
     completion: @escaping (PaywallViewController?, PaywallSkippedReason?, Error?) -> Void
   ) {
     Task { @MainActor in
@@ -47,7 +40,8 @@ extension Superwall {
         let paywallViewController = try await getPaywallViewController(
           forEvent: event,
           params: params,
-          paywallOverrides: paywallOverrides
+          paywallOverrides: paywallOverrides,
+          delegate: delegate
         )
         completion(paywallViewController, nil, nil)
       } catch let reason as PaywallSkippedReason {
@@ -61,15 +55,6 @@ extension Superwall {
   /// Gets the  ``PaywallViewController`` object, which you can present
   /// however you want.
   ///
-  /// To use this you **must** follow the following steps:
-  ///
-  /// 1. Call this function to retrieve the ``PaywallViewController``.
-  /// 2. Call ``PaywallViewController/presentationWillBegin()`` when
-  /// you're about to present the view controller.
-  /// 3. Present the view controller.
-  /// 4. Call ``PaywallViewController/presentationDidFinish()`` after presentation
-  /// completes.
-  ///
   /// - Note: The remotely configured presentation style will be ignored, it is up to you
   /// to set it programmatically.
   ///
@@ -80,36 +65,32 @@ extension Superwall {
   ///   JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will
   ///   be dropped.
   ///   - paywallOverrides: An optional ``PaywallOverrides`` object whose parameters override the paywall defaults. Use this to override products, presentation style, and whether it ignores the subscription status. Defaults to `nil`.
+  ///   - delegate: A delegate responsible for handling user interactions with the retrieved ``PaywallViewController``.
   ///
-  /// - Returns A ``PaywallViewController`` object.
+  /// - Returns: A ``PaywallViewController`` object.
   /// - Throws: An `Error` explaining why it couldn't get the view controller. If the ``PaywallViewController`` couldn't be retrieved
-  /// because its presentation should be skipped, catch an error of type``PaywallSkippedReason`` and switch over its cases to find out
+  /// because its presentation should be skipped, catch an error of type ``PaywallSkippedReason`` and switch over its cases to find out
   /// more info. All other errors will be returned in the general catch block.
   @MainActor
   @nonobjc public func getPaywallViewController(
     forEvent event: String,
     params: [String: Any]? = nil,
-    paywallOverrides: PaywallOverrides? = nil
+    paywallOverrides: PaywallOverrides? = nil,
+    delegate: PaywallViewControllerDelegate
   ) async throws -> PaywallViewController {
     return try await internallyGetPaywallViewController(
       forEvent: event,
       params: params,
       paywallOverrides: paywallOverrides,
-      isObjc: false
+      delegate: .init(
+        swiftDelegate: delegate,
+        objcDelegate: nil
+      )
     )
   }
 
   /// Objective-C-only method that gets the  ``PaywallViewController`` object, which you can present
   /// however you want.
-  ///
-  /// To use this you **must** follow the following steps:
-  ///
-  /// 1. Call this function to retrieve the ``PaywallViewController``.
-  /// 2. Call ``PaywallViewController/presentationWillBegin()`` when
-  /// you're about to present the view controller.
-  /// 3. Present the view controller.
-  /// 4. Call ``PaywallViewController/presentationDidFinish()`` after presentation
-  /// completes.
   ///
   /// - Note: The remotely configured presentation style will be ignored, it is up to you
   /// to set it programmatically.
@@ -122,13 +103,17 @@ extension Superwall {
   ///   be dropped.
   ///   - paywallOverrides: An optional ``PaywallOverrides`` object whose parameters override the paywall
   ///   defaults. Use this to override products, presentation style, and whether it ignores the subscription status. Defaults to `nil`.
-  ///   - completion: A completion block that accepts a ``GetPaywallViewControllerResult`` object. First check ``GetPaywallViewControllerResult/paywallViewController`` to see if was retrieved. Then check ``GetPaywallViewControllerResult/skippedReason`` to see if it's presentation was intentionally skipped. Then check
+  ///   - completion: A completion block that accepts a ``GetPaywallViewControllerResult`` object. First check
+  ///   ``GetPaywallViewControllerResult/paywallViewController`` to see if was retrieved. Then check
+  ///   ``GetPaywallViewControllerResult/skippedReason`` to see if it's presentation was intentionally skipped. Then check
   ///   ``GetPaywallViewControllerResult/error`` for any errors that may have occurred.
+  ///   - delegate: A delegate responsible for handling user interactions with the retrieved ``PaywallViewController``.
   @available(swift, obsoleted: 1.0)
   @objc public func getPaywallViewController(
     forEvent event: String,
     params: [String: Any]? = nil,
     paywallOverrides: PaywallOverrides? = nil,
+    delegate: PaywallViewControllerDelegateObjc,
     completion: @escaping (GetPaywallViewControllerResult) -> Void
   ) {
     Task { @MainActor in
@@ -137,7 +122,10 @@ extension Superwall {
           forEvent: event,
           params: params,
           paywallOverrides: paywallOverrides,
-          isObjc: true
+          delegate: .init(
+            swiftDelegate: nil,
+            objcDelegate: delegate
+          )
         )
         let reason = GetPaywallViewControllerResult(
           paywallViewController: paywallViewController,
@@ -167,7 +155,7 @@ extension Superwall {
     forEvent event: String,
     params: [String: Any]?,
     paywallOverrides: PaywallOverrides?,
-    isObjc: Bool
+    delegate: PaywallViewControllerDelegateAdapter
   ) async throws -> PaywallViewController {
     return try await Future {
       let trackableEvent = UserInitiatedEvent.Track(
@@ -184,9 +172,9 @@ extension Superwall {
         .explicitTrigger(trackResult.data),
         paywallOverrides: paywallOverrides,
         isPaywallPresented: false,
-        type: .getPaywallViewController
+        type: .getPaywallViewController(delegate)
       )
-      return self.getPaywallViewController(presentationRequest, isObjc: isObjc)
+      return self.getPaywallViewController(presentationRequest)
     }
     .eraseToAnyPublisher()
     .throwableAsync()