+++ Sources/SuperwallKit/Paywall/Presentation/Internal Presentation/Presentation State/PaywallState.swift
@@ -15,18 +15,34 @@ public enum PaywallResult: Equatable, Sendable {
   ///   - productId: The identifier of the product.
   case purchased(productId: String)
 
-  /// The paywall was dismissed by the user manually pressing the close button.
-  case closed
+  /// The paywall was declined by the user pressing the close button.
+  case declined
 
   /// The paywall was dismissed due to the user restoring their purchases.
   case restored
+
+  func convertForObjc() -> PaywallResultObjc {
+    switch self {
+    case .purchased:
+      return .purchased
+    case .declined:
+      return .declined
+    case .restored:
+      return .restored
+    }
+  }
 }
 
-/// Objective-C compatible enum for ``PaywallResult``
+/// Objective-C-only enum. Contains the possible reasons for the dismissal of a paywall.
 @objc(SWKPaywallResult)
 public enum PaywallResultObjc: Int, Sendable {
+  /// The paywall was dismissed because the user purchased a product
   case purchased
-  case closed
+
+  /// The paywall was declined by the user pressing the close button.
+  case declined
+
+  /// The paywall was dismissed due to the user restoring their purchases.
   case restored
 }