+++ CHANGELOG.md
@@ -6,11 +6,20 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 
 ### Breaking Changes
 
-- `getPaywallViewController(forEvent:params:paywallOverrides:completion:)` now accepts an optional `PaywallViewController`, an optional `PaywallSkippedReason` and an optional `Error`. This makes it easier to understand when the paywall was skipped vs when a real error occurred.
+- Adds a `PaywallViewControllerDelegate` to the `getPaywallViewController` functions. This is mandatory and is how you control what happens after a paywall is dismissed.
+- The completion block of `getPaywallViewController(forEvent:params:paywallOverrides:delegate:completion:)` now accepts an optional `PaywallViewController`, an optional `PaywallSkippedReason` and an optional `Error`. This makes it easier to understand when the paywall was skipped vs when a real error occurred.
+- Renamed the `PaywallResult` case `closed` to `declined`.
 
 ### Enhancements
 
 - Exposes `PaywallOverrides` and `PaywallViewController` to Objective-C.
+- Adds Objective-C convenience methods to `PaywallOverrides`.
+- Adds a `device.isFirstSession` property that you can use in paywall rules. This is `true` for the very first time a user opens the app. When the user closes and reopens the app, this will be `false`.
+- Removes the need to tell us when you're going to present/have presented a `PaywallViewController` that has been retrieved using `getPaywallViewController(...)`.
+
+### Fixes
+
+- Fixes a crash when calling `reset()` when a paywall is displayed.
 
 ## 3.0.0-rc.5